Add a test to force the tree to close
Right now we don't have a good way to propogate that the tree is closed.
Add a test that can be manually changed to fail to turn the tree red.
Also starts off with the assert failing since the tree should be red
right now.

Long term something like this would be controlled by a better mechanism,
but this works for now.